+++ packages/javascript/lib/web/index.ts
@@ -1,4 +1,5 @@
 export { generateLicense } from './generateLicense.js';
+export { getParsedInfoFromAddress } from './getParsedInfoFromAddress.js';
 export { isBotAgent } from './isBotAgent.js';
 export { isMatchPathname } from './isMatchPathname.js';
 export { isMobile } from './isMobile.js';